feat(core-banking): API for test
- Implemented test methods

+++ spring-consumer/src/main/java/org/consumer/repository/DelayedTransferRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DelayedTransferRepository extends JpaRepository<DelayedTransferRequest, Long> {
+public interface DelayedTransferRepository extends JpaRepository<DelayedTransferRequest, Integer> {
 }
